Worker container Switched Makefile -> DockerHub
We want to Autobuild our containers. Therefore we need to move the functionally
of the Makefiles inside our Dockerfiles. So I added git to the packages to be
installed and pull the last TBro version from GitHub during container
build. Copying the required file to the expected location is the last step.