ChadoReload container Switched Makefile -> DockerHub
We want to Autobuild our containers. Therefore we need to move the functionally
of the Makefiles inside our Dockerfiles. In the case of the ChadoReload
container I removed the Makefile completly. Furthermore I linked the current
chado database dump into the Dockerfile directory and changed the names to be
more flexible updating the chado database dump.