tbro_apache container Switched Makefile -> DockerHub
We want to Autobuild our containers. Therefore we need to move the functionally
of the Makefiles inside our Dockerfiles. In the case of the tbro_apache
container I removed the Makefile completly. Furthermore the current tbro GitHub
repository will be cloned and its content will be copied to the folder
/home/tbro.