tbro_demo container Switched Makefile -> DockerHub
We want to Autobuild our containers. Therefore we need to move the functionally
of the Makefiles inside our Dockerfiles. In the case of the tbro_demo container
I removed the Makefile completly. Furthermore the current tbro DemoData GitHub
repository will be cloned and the content of its master branch will be used to
import the demo data.